// types/transaction.ts
export enum TRANSACTION_TYPES {
  SET_DATA = 'SET_DATA',
  SET_DATA_BATCH = 'SET_DATA_BATCH',
  EXECUTE = 'EXECUTE',
  EXECUTE_BATCH = 'EXECUTE_BATCH',
  TRANSFER_OWNERSHIP = 'TRANSFER_OWNERSHIP',
  ACCEPT_OWNERSHIP = 'ACCEPT_OWNERSHIP',
  RENOUNCE_OWNERSHIP = 'RENOUNCE_OWNERSHIP',
  BATCH_CALLS = 'BATCH_CALLS',
  UNIVERSAL_RECEIVER = 'UNIVERSAL_RECEIVER',
}